Precomputed pdf links bug fix.

+++ src/vpfHeader1General0_General.h
@@ -1112,7 +1112,7 @@ public:
   }
   inline Object& operator[](int i)const
   { this->CheckConsistency();
-    if ((i>=this->size) || (i<0))
+    if (((int)i >= (int) this->size) || ((int) i < 0))
       crash << "Programming error: attempting to access the entry of index " << i << " in an array of " << this->size << " elements. " << crash;
     return this->TheObjects[i];
   }
